Add: SliverList. Add: GestureDetector.
Add SliverList in pelicula detalle to give body to the page, including the movie title in the poster, the movie title and the original title in english and the rate of the movie.
Add GestureDetector in movie horizontal to include the arguments of the class pelicula and send it to another page as pelicula detalle, once the user tap in any of the popular movie.