Add: Unique class, add: Gesture Detector.
Add unique id class in pelicula model to avoid the aplication to show pelicula detalle even if the movie is in the popular movies and in the card swiper.
Add Gesture detector in the card swiper to allow the aplication to show the pelicula detalle when the user tap in a movie card.